refactor(runtime): remove retired process-containment alias tombstones
Delete retired process-containment environment identifiers and migration-only fixtures while preserving canonical trusted boundaries.
